[backend] refactor: mutualize utils files
