+++ opencti-platform/opencti-graphql/src/modules/case/case-incident/case-incident-converter.ts
@@ -1,7 +1,8 @@
-import { buildStixDomain, cleanObject } from '../../../database/stix-2-1-converter';
+import { buildStixDomain } from '../../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../../types/stix-2-1-extensions';
 import { INPUT_OBJECTS } from '../../../schema/general';
 import type { StixCaseIncident, StoreEntityCaseIncident } from './case-incident-types';
+import { cleanObject } from '../../../database/stix-converter-utils';
 
 const convertCaseIncidentToStix = (instance: StoreEntityCaseIncident): StixCaseIncident => {
   const caseIncident = buildStixDomain(instance);
