+++ opencti-platform/opencti-graphql/src/modules/case/case-rfi/case-rfi-converter.ts
@@ -1,7 +1,8 @@
-import { buildStixDomain, cleanObject } from '../../../database/stix-2-1-converter';
+import { buildStixDomain } from '../../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../../types/stix-2-1-extensions';
 import { INPUT_OBJECTS } from '../../../schema/general';
 import type { StixCaseRfi, StoreEntityCaseRfi } from './case-rfi-types';
+import { cleanObject } from '../../../database/stix-converter-utils';
 
 const convertCaseRfiToStix = (instance: StoreEntityCaseRfi): StixCaseRfi => {
   const caseRfi = buildStixDomain(instance);
