+++ opencti-platform/opencti-graphql/src/modules/case/case-rft/case-rft-converter.ts
@@ -1,7 +1,8 @@
-import { buildStixDomain, cleanObject } from '../../../database/stix-2-1-converter';
+import { buildStixDomain } from '../../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../../types/stix-2-1-extensions';
 import { INPUT_OBJECTS } from '../../../schema/general';
 import type { StixCaseRft, StoreEntityCaseRft } from './case-rft-types';
+import { cleanObject } from '../../../database/stix-converter-utils';
 
 const convertCaseRftToStix = (instance: StoreEntityCaseRft): StixCaseRft => {
   const caseRft = buildStixDomain(instance);
