+++ opencti-platform/opencti-graphql/src/modules/case/case-template/case-template-converter.ts
@@ -1,6 +1,7 @@
-import { buildStixDomain, cleanObject } from '../../../database/stix-2-1-converter';
+import { buildStixDomain } from '../../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../../types/stix-2-1-extensions';
 import type { StixCaseTemplate, StoreEntityCaseTemplate } from './case-template-types';
+import { cleanObject } from '../../../database/stix-converter-utils';
 
 const convertCaseTemplateToStix = (instance: StoreEntityCaseTemplate): StixCaseTemplate => {
   const caseTemplate = buildStixDomain(instance);
