+++ opencti-platform/opencti-graphql/src/modules/dataSource/dataSource-converter.ts
@@ -1,6 +1,7 @@
 import type { StixDataSource, StoreEntityDataSource } from './dataSource-types';
-import { buildStixDomain, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixDomain } from '../../database/stix-2-1-converter';
 import { STIX_EXT_MITRE, STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertDataSourceToStix = (instance: StoreEntityDataSource): StixDataSource => {
   const stixDomainObject = buildStixDomain(instance);
