+++ opencti-platform/opencti-graphql/src/modules/decayRule/decayRule-converter.ts
@@ -1,6 +1,7 @@
-import { buildStixObject, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixObject } from '../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
 import type { StixDecayRule, StoreEntityDecayRule } from './decayRule-types';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertDecayRuleToStix = (instance: StoreEntityDecayRule): StixDecayRule => {
   const stixObject = buildStixObject(instance);
