+++ opencti-platform/opencti-graphql/src/modules/disseminationList/disseminationList-converter.ts
@@ -13,9 +13,10 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 */
 
-import { buildStixObject, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixObject } from '../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
 import type { StixDisseminationList, StoreEntityDisseminationList } from './disseminationList-types';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertDisseminationListToStix = (instance: StoreEntityDisseminationList): StixDisseminationList => {
   const stixObject = buildStixObject(instance);
