+++ opencti-platform/opencti-graphql/src/modules/exclusionList/exclusionList-converter.ts
@@ -1,6 +1,7 @@
 import type { StixExclusionList, StoreEntityExclusionList } from './exclusionList-types';
-import { buildStixObject, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixObject } from '../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertExclusionListToStix = (instance: StoreEntityExclusionList): StixExclusionList => {
   const stixObject = buildStixObject(instance);
