+++ opencti-platform/opencti-graphql/src/modules/fintelTemplate/fintelTemplate-converter.ts
@@ -1,6 +1,7 @@
-import { buildStixObject, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixObject } from '../../database/stix-2-1-converter';
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
 import type { StixFintelTemplate, StoreEntityFintelTemplate } from './fintelTemplate-types';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 export const convertFintelTemplateToStix = (instance: StoreEntityFintelTemplate): StixFintelTemplate => {
   const stixObject = buildStixObject(instance);
