+++ opencti-platform/opencti-graphql/src/modules/grouping/grouping-converter.ts
@@ -1,7 +1,8 @@
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
-import { buildStixDomain, cleanObject, convertObjectReferences } from '../../database/stix-2-1-converter';
+import { buildStixDomain, convertObjectReferences } from '../../database/stix-2-1-converter';
 import type { StixGrouping, StoreEntityGrouping } from './grouping-types';
 import { INPUT_OBJECTS } from '../../schema/general';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertGroupingToStix = (instance: StoreEntityGrouping): StixGrouping => {
   const grouping = buildStixDomain(instance);
