+++ opencti-platform/opencti-graphql/src/modules/internal/csvMapper/csvMapper-converter.ts
@@ -1,6 +1,7 @@
-import { buildStixObject, cleanObject } from '../../../database/stix-2-1-converter';
+import { buildStixObject } from '../../../database/stix-2-1-converter';
 import type { StixCsvMapper, StoreEntityCsvMapper } from './csvMapper-types';
 import { STIX_EXT_OCTI } from '../../../types/stix-2-1-extensions';
+import { cleanObject } from '../../../database/stix-converter-utils';
 
 const convertCsvMapperToStix = (instance: StoreEntityCsvMapper): StixCsvMapper => {
   const stixObject = buildStixObject(instance);
