+++ opencti-platform/opencti-graphql/src/modules/language/language-converter.ts
@@ -1,7 +1,8 @@
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
-import { buildStixObject, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixObject } from '../../database/stix-2-1-converter';
 import type { StixLanguage, StoreEntityLanguage } from './language-types';
 import { INPUT_CREATED_BY } from '../../schema/general';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertLanguageToStix = (instance: StoreEntityLanguage): StixLanguage => {
   const stixDomainObject = buildStixObject(instance);
