+++ opencti-platform/opencti-graphql/src/modules/managerConfiguration/managerConfiguration-converter.ts
@@ -1,6 +1,7 @@
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
-import { buildStixObject, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixObject } from '../../database/stix-2-1-converter';
 import type { StixManagerConfiguration, StoreEntityManagerConfiguration } from './managerConfiguration-types';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertManagerConfigurationToStix = (instance: StoreEntityManagerConfiguration): StixManagerConfiguration => {
   const stixObject = buildStixObject(instance);
