+++ opencti-platform/opencti-graphql/src/modules/narrative/narrative-converter.ts
@@ -1,6 +1,7 @@
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
-import { buildStixDomain, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixDomain } from '../../database/stix-2-1-converter';
 import type { StixNarrative, StoreEntityNarrative } from './narrative-types';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertNarrativeToStix = (instance: StoreEntityNarrative): StixNarrative => {
   const stixDomainObject = buildStixDomain(instance);
