+++ opencti-platform/opencti-graphql/src/modules/playbook/playbook-converter.ts
@@ -14,8 +14,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 */
 
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
-import { buildStixObject, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixObject } from '../../database/stix-2-1-converter';
 import type { StixPlaybook, StoreEntityPlaybook } from './playbook-types';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertEntityPlaybookToStix = (instance: StoreEntityPlaybook): StixPlaybook => {
   const stixObject = buildStixObject(instance);
