+++ opencti-platform/opencti-graphql/src/modules/publicDashboard/publicDashboard-converter.ts
@@ -1,6 +1,7 @@
 import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
-import { buildStixDomain, cleanObject } from '../../database/stix-2-1-converter';
+import { buildStixDomain } from '../../database/stix-2-1-converter';
 import type { StixPublicDashboard, StoreEntityPublicDashboard } from './publicDashboard-types';
+import { cleanObject } from '../../database/stix-converter-utils';
 
 const convertPublicDashboardToStix = (instance: StoreEntityPublicDashboard): StixPublicDashboard => {
   const stixDomainObject = buildStixDomain(instance);
